pyproject.toml
[project]
name = "prompt-runner"
version = "0.1.0"
description = "A multi-turn LLM conversation agent with dynamic tool support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "litellm[google]>=1.40.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]
